plugins: Add pre_ and post_ hooks to core functions
If a loaded plugin defines a pre_/post_{fname} function, it will be run
before/after that function. For example, if a plugin defines post_run,
that function will be executed after a 'pman run' via the CLI or a
pman.run() via the API.

These hooks take 'config' as the first argument.