build: Run build tasks in parallel
The number of concurrent jobs is controlled by "jobs" under "general".
If set to 0, this will allow a number of jobs equal to the number of
processors on the system. The default is 0.

Also use Rich to make build output prettier and more readable.

Batching (i.e., having blend2bam process multiple
blend files at once) is currently forced to 1. This will result in
slower builds if the overhead of launching Blender for each blend file
is still larger than the amount of time saved by running multiple
instances of Blender.