build: Add support for defining explicit streams
If the new build.streams config is left empty, the old behavior of
auto-detecting plugins to use to convert files is used.

If the list is non-empty, each item specifies a plugin,
include/exclude_patterns, and plugin options used to determine
(explicitly) which files to build and how to build them.